Simple Gaming Social Network
You and your friend have decided to start a company that hosts a gaming
social network site. Your friend will handle the website creation (they
know what they are doing, having taken our web development class).
However, it is up to you to create a data structure that manages the
game-network information and to define several procedures that operate
on the network.